總倉月結算明細整頁化:去彈窗+頁內確認+調撥單連結
跟進分店端整頁化，總倉明細也改整頁：
- 新頁 /transfers/settlement/detail?id=，列表「明細」直接連過去；
  統計卡/流程狀態列/爭議面板/改估價行內面板照搬。
- 動作全改頁內確認、不跳瀏覽器對話框：送店家核對/直接確認/
  收款結案＝底部固定確認列；處理爭議＝爭議列行內展開說明輸入
  （取代 window.prompt）。
- 明細調撥單號可點（/wms/transfers?open= 深連結），與分店端一致。
- 移除列表頁的明細 Modal 與整個 HqToStoreDetail component。

Playwright 12/12 通過（含監聽 dialog 驗證 0 彈窗）。
